requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
python-multipart 
scikit-learn
